[Silver III] Title: 주유소, Time: 336 ms, Memory: 40644 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -30,8 +30,8 @@ public class Main {
 //        System.out.println();
 
         long totalPrice= 0;
-        int distance = distances[0];
-        int minPrice = prices[0];
+        long distance = distances[0];
+        long minPrice = prices[0];
 
         for(int i = 1; i < N - 1; i++) {
             if(minPrice > prices[i]) {
